requirements.txt
pymupdf
ebooklib
beautifulsoup4
pyyaml
pymupdf4llm
pandas